pyproject.toml
[project]
name = "backtask"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.1",
    "psycopg2-binary>=2.9.11",
    "sqlalchemy>=2.0.48",
    "uvicorn>=0.41.0",
    "pydantic>=2.0",
    "python-jose>=3.3.0",
    "passlib>=1.7.4",
    "bcrypt>=4.1.0",
    "python-multipart>=0.0.6",
    "httpx>=0.27.0",
    "pytest>=8.0.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]